Changes on SmartMonitor interface: Replaced QueueStatus (ocupation, threshold, dropProb, maxThreshold) with simply dropProb. Updated BE, DL and eDL with those changes
Changed DumbMaxQ, it simply ask for dropProb and decides given that
Removed PortMaxQ as was almost the same as DumbMaxQ

First version of DQ policy.
DQSched similar to DumbSched, but on input it ask for nextPDU time if no queue returned.
DQMonitor, all computatios of DQ. Configuration as:
Limit (input). Limit data include info about PDU spacing and burst drop. Limits imposed to all the input queues of a port, so if Q1 and Q2 have the same Limit, (e.j. 10Mbps), Q1+Q2 cannot surpase that Limit.
Cherish (output). Defines the probability of drop a PDU depending on the number of PDUs waiting on port.
Urgency (output). Used to decide the urgency of a new PDU. Degradation can happen depending on the rate on the queue.
Clases are stored as a tuple <queueName, Limit, Cherish, Urgency>